Move fuzzing into nocover
