+++ tests/nocover/test_regex.py
@@ -21,7 +21,8 @@ import re
 import string
 
 import hypothesis.strategies as st
-from hypothesis import given, assume
+from hypothesis import given, assume, reject
+from hypothesis.errors import NoExamples, FailedHealthCheck
 from hypothesis.searchstrategy.regex import base_regex_strategy
 
 
@@ -63,3 +64,20 @@ def test_conservative_regex_are_correct_by_construction(data):
     pattern = re.compile(pattern)
     result = data.draw(base_regex_strategy(pattern))
     assert pattern.match(result) is not None
+
+
+@given(st.text(max_size=100) | st.binary(max_size=100) | CONSERVATIVE_REGEX)
+def test_fuzz_stuff(pattern):
+    try:
+        regex = re.compile(pattern)
+    except re.error:
+        reject()
+
+    @given(st.from_regex(regex))
+    def inner(ex):
+        assert regex.match(ex)
+
+    try:
+        inner()
+    except (NoExamples, FailedHealthCheck):
+        reject()
